requirements.txt
Langchain
ipykernel
python-dotenv
Langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
duckdb
duckduckgo-search
sqlalchemy
mysql-connector-python
validators==0.28.1
youtube_transcript_api
pytube
unstructured
huggingface_hub
openai
langchain_nvidia_ai_endpoints
